setup/tooling: skill-apply heading-ordinal strip, headless browser URL surfacing, inherit-script extraction
Three small wizard/engine fixes: strip leading authoring ordinals from
skill-apply step captions, always print the raw auth URL on headless
installs, and extract runInheritScript into setup/lib so channel flows
can reuse it without importing the wizard driver.
