+++ setup/lib/browser.ts
@@ -71,7 +71,15 @@ export async function confirmThenOpen(
   url: string,
   message = 'Press Enter to open your browser',
 ): Promise<void> {
-  if (isHeadless()) return;
+  if (isHeadless()) {
+    // No browser to open here — ALWAYS surface the URL itself, and print it
+    // RAW on stdout: clack notes wrap long lines and inject gutter pipes and
+    // spaces mid-URL, breaking copy-paste (live-hit). Plain stdout is the
+    // only rendering that keeps the URL contiguous.
+    p.log.info('Open this URL in a browser on any device:');
+    process.stdout.write(`\n${url}\n\n`);
+    return;
+  }
   const fallback = `\n${k.dim(`If browser does not appear, please visit: ${url}`)}`;
   ensureAnswer(
     await p.confirm({
